Go back to manually starting new processes
This fixes an issue for me where I would not be able to get a full list of available
subroutines even though I should be able to. There is a slight performance hit
from starting a new process for every completion, but it's worth it.

Also added code to look at parent classes for their functions as well.